adding mailer.js util and framework for email firing
